Move models.Logs to application.Log
